team_23-main/requirements.txt
black==24.8.0
dash_bootstrap_components==1.6.0
dash-html-components==2.0.0
dash-table==5.0.0
dash==2.18.1
db-dtypes==1.3.0
Flask==3.0.3
functions-framework>=3.4.0
google-cloud-bigquery
gunicorn==23.0.0
lightgbm==4.3.0
loguru==0.7.2
matplotlib==3.9.2
nba_api==1.5.2
numpy==1.26.4
pandas==2.2.2
plotly==5.15.0
protobuf==4.25.5
Requests==2.32.3
scikit_learn==1.4.0
setuptools>=69.2.0
sqlfluff==3.2.0
statsmodels==0.14.1
dbt-core==1.8.8
dbt-bigquery==1.8.3
seaborn==0.13.2